+++ CHANGES.rst
@@ -17,6 +17,8 @@ Version 0.9.1 (UNRELEASED)
     - Fixes cronjob failures due to database connection issues when REANA is deployed with non-default namespace or prefix.
     - Fixes ``ingress.enabled`` option to correctly enable or disable the creation of Ingresses.
     - Adds new configuration options ``login`` and ``secrets.login`` for configuring Keycloak SSO login with third-party authentication services.
+- Developers:
+    - Fixes ``cluster-deploy`` and ``cluster-undeploy`` commands when using custom instance name or kubernetes namespace.
 
 Version 0.9.0 (2023-01-26)
 --------------------------